Adapt azure and google cloud storage source into tiered tsts

+++ test/suite/test_tiered06.py
@@ -66,7 +66,7 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
             conf += ',cache_directory=' + cache_dir
         return conf
 
-    def test_ss_basic(self):
+    def ss_basic(self):
         # Test some basic functionality of the storage source API, calling
         # each supported method in the API at least once.
 
@@ -137,7 +137,7 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
                 lambda: fs.fs_rename(session, 'foobar', 'barfoo', 0))
         self.assertEquals(fs.fs_directory_list(session, '', ''), ['foobar'])
 
-        if self.ss_name == 's3_store':
+        if self.ss_name != 'dir_store':
             # Files that have been flushed cannot be manipulated through the custom file system.
             with self.expectedStderrPattern('foobar: remove of file not supported'):
                 self.assertRaisesException(wiredtiger.WiredTigerError,
@@ -150,7 +150,7 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
         fs.terminate(session)
         ss.terminate(session)
 
-    def test_ss_write_read(self):
+    def ss_write_read(self):
         # Write and read to a file non-sequentially.
 
         session = self.session
@@ -232,7 +232,8 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
                 else:
                     self.assertEquals(in_block, a_block)
             fh.close(session)
-            os.remove(os.path.join(cachedir, 'abc'))
+            if (self.ss_name != "gcp_store" and self.ss_name != "azure_store"):
+                os.remove(os.path.join(cachedir, 'abc'))
 
         ss.terminate(session)
 
@@ -312,10 +313,11 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
         bad_cachedir = '/BAD'
 
         # Create file system objects. First try some error cases.
-        errmsg = '/No such /'
-        self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
-            lambda: ss.ss_customize_file_system(session, self.bucket, self.auth_token,
-                self.get_fs_config(prefix, bad_cachedir)), errmsg)
+        errmsg = '/No such|Invalid bucket name/'
+        if self.has_cache:
+            self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
+                lambda: ss.ss_customize_file_system(session, self.bucket, self.auth_token,
+                    self.get_fs_config(prefix, bad_cachedir)), errmsg)
 
         # S3 store expects a region with the bucket
         if self.ss_name == 's3_store':
@@ -352,7 +354,8 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
         self.check_home(['beagle', 'bird', 'bison', 'bat', 'cat', 'cougar', 'coyote', 'cub'])
         self.check_dirlist(fs1, '', [])
         self.check_dirlist(fs2, '', [])
-        self.check_caches([], [])
+        if (self.has_cache):
+            self.check_caches([], [])
         self.check_local_objects([], [])
 
         # A flush copies to the cloud, nothing is removed.
@@ -360,7 +363,8 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
         self.check_home(['beagle', 'bird', 'bison', 'bat', 'cat', 'cougar', 'coyote', 'cub'])
         self.check_dirlist(fs1, '', ['beagle'])
         self.check_dirlist(fs2, '', [])
-        self.check_caches([], [])
+        if (self.has_cache):
+            self.check_caches([], [])
         self.check_local_objects(['beagle'], [])
 
         # Bad file to flush.
@@ -373,7 +377,8 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
         self.check_home(['beagle', 'bird', 'bison', 'bat', 'cat', 'cougar', 'coyote', 'cub'])
         self.check_dirlist(fs1, '', ['beagle'])
         self.check_dirlist(fs2, '', [])
-        self.check_caches([], [])
+        if (self.has_cache):
+            self.check_caches([], [])
         self.check_local_objects(['beagle'], [])
 
         # When we flush_finish, the local file will be in both the local and cache directory.
@@ -381,7 +386,8 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
         self.check_home(['beagle', 'bird', 'bison', 'bat', 'cat', 'cougar', 'coyote', 'cub'])
         self.check_dirlist(fs1, '', ['beagle'])
         self.check_dirlist(fs2, '', [])
-        self.check_caches(['beagle'], [])
+        if (self.has_cache):
+            self.check_caches(['beagle'], [])
         self.check_local_objects(['beagle'], [])
 
         # Do a some more in each file system.
@@ -395,7 +401,8 @@ class test_tiered06(wttest.WiredTigerTestCase, TieredConfigMixin):
         self.check_home(['beagle', 'bird', 'bison', 'bat', 'cat', 'cougar', 'coyote', 'cub'])
         self.check_dirlist(fs1, '', ['beagle', 'bat', 'bison'])
         self.check_dirlist(fs2, '', ['cat', 'cub'])
-        self.check_caches(['beagle', 'bat'], ['cat'])
+        if (self.has_cache):
+            self.check_caches(['beagle', 'bat'], ['cat'])
         self.check_local_objects(['beagle', 'bat', 'bison'], ['cat', 'cub'])
 
         # Test directory listing prefixes.
